user-auth-service/requirements.txt
fastapi
uvicorn
python-dotenv
pytz
pyjwt
python-jose
requests
ldap3
pydantic
python-jose[cryptography]
passlib
python-dotenv
cryptography
pycryptodome
starlette
prometheus_fastapi_instrumentator
slowapi
sqlalchemy
pandas
openpyxl
psycopg2